Fixed bug in listp which generates segfault
listp returns t if its parameters is a list, so it is necessary run over
the cdr of the object until cdr becomes nil or some car isn't a atom,
because in this case means we have a tree and not a list.

Code ran over the list without testing nil before, so in case parameter was
nil or a atom t generates a segfault. For getting rid of this is added a
previous testing.

There is also other bug because condition in the loop was while !ATOMP(obj)
and the correct is the inverse, ATOMP(obj)