Fixed bug in minus operation
kemacs doesn't support negative numbers, so it is necessary check this
condition in the minus operation, so we only accept a minimun of 2 parameters
because in other case result will be negative.

Other problem with the previous code was that it increments the pointer to
args, and this is an error, because it is a stack, so it is necessary
decrement it, and also use XINT before of assigning the value to cnt
variable, which was type l_object and this is a unsigned type, so the
negative check was always false.