Added EOF handling
If we get an EOF while you are parsing an sexp is needed raise an
unrecoverable error. If we get the EOF outside of an sexp is valid,
because this means we reached end of file while we were searching a
new sexp, so it is correct.

Due to this a new error handling is added, eof_while_parsing that has
different semantic that end_of_file has.