Added support for reading cons cells
Cons cell have diferent syntax because they are read as (1 . 2)
instead of (1 2) how it is in list case.

It is needed modify readlist in order to read dot in the correct
context, so after a dot always must come a new element and later
')'. In other case there is an error.